When they release a new device profile, is testing done?
Based on the common configuration that was forked from CM9 to allow
easier combination of similar devices, the build would not boot.
Multiple components were simply mirrored from msm7x30, which if
possible would void the need to create a secondary repository. This is
part of a collection of commits that bring it back into working order.